Update bcryptjs to 3.0.3 and remove setInterval workaround
