+++ package.json
@@ -101,7 +101,7 @@
     "ansible-vault": "^1.3.0",
     "apache-md5": "^1.1.8",
     "arr-diff": "^4.0.0",
-    "bcryptjs": "^2.4.3",
+    "bcryptjs": "^3.0.3",
     "big.js": "^6.2.2",
     "braces": "^3.0.3",
     "bwip-js": "^4.5.1",
